+++ js/copy/removeCopy.mjs
@@ -1,0 +1,7 @@
+function removeCopy() {
+  enhancedElements2.forEach(ee => {
+    ee.element.removeEventListener("click", ee.copied);
+  });
+}
+
+export default removeCopy;
