+++ js/settings/copySetting.js
@@ -2,6 +2,7 @@ import Doc from "../service/doc.mjs";
 import { copySetting } from "./localVariable.mjs";
 import removeAutoCopy from "../copy/removeAutoCopy.mjs";
 import autocopy from "../copy/autocopy.mjs";
+import onecopy from "../copy/onecopy.mjs";
 
 const copybtn = Doc.findAll(".autocopyBtn");
 
@@ -39,7 +40,7 @@ function addCopyEvent() {
     copybtn[1].classList.add("active");
     await btnReset();
     await removeAutoCopy();
-    // await onecopy();
+    await onecopy();
   });
 }
